+++ test/obiwan/test/tools.clj
@@ -1,7 +1,10 @@
 (ns obiwan.test.tools
   (:require [obiwan.core :as redis]
+            [obiwan.tools :as t]
+            [clojure.java.io :as io]
             [yang.lang :as y])
   (:import [redis.embedded RedisServer
+                           RedisServerBuilder
                            RedisExecProvider]
            [redis.embedded.util OS]))
 
@@ -9,15 +12,22 @@
                                  os
                                  architecture
                                  port
-                                 modules] ;; path-to-server --loadmodule path-to/redisearch.so
-                          :or {os "MAC_OS_X"}
+                                 dir           ;; TODO: convert to a settings map
+                                 modules]      ;; path-to-server --loadmodule path-to/redisearch.so
+                          :or {os "MAC_OS_X"
+                               dir "/tmp"}
                           :as opts}]
   (println "making redis server" opts)
   (let [provider (doto (RedisExecProvider/defaultProvider)
-                       (.override (OS/MAC_OS_X/valueOf os)
-                                  path))]
-
-    (RedisServer. provider port)))
+                   (.override (OS/MAC_OS_X/valueOf os)
+                              path))
+        rdbfile (str dir "/dump.rdb")]
+    (io/delete-file rdbfile true)
+    (-> (RedisServerBuilder.)
+        (.redisExecProvider provider)
+        (.port port)
+        (.setting (str "dir " dir))
+        (.build))))
 
 (defn start-redis-server
   ([]
@@ -61,3 +71,24 @@
 (defn with-flushall [f]
     (f)
     (redis/say conn "FLUSHALL"))
+
+(defn load-module
+  ([conn mname]
+   (load-module conn mname {}))
+  ([conn
+    mname
+    {:keys [config]
+     :or {config "resources/config.edn"}}]
+   (let [module (-> (y/edn-resource config)
+                    :redis
+                    :server
+                    :modules
+                    mname)]
+     (println "loading module" {mname module})
+     (redis/module-load conn module))))
+
+(defn with-search-module [f]
+  (load-module conn :search)
+  (f)
+  ;; (redis/module-unload conn "search")  ;; can't unload since the module introduces a new data structure
+  )